# Optimize memory allocation in histogram iterators